requirements.txt
pillow
torch
torchvision
scipy
sendgrid
scikit-image
gdown
requests
opencv-python
opencv-contrib-python
scipy
requests
opencv-python
flask_cors